[Rx]: Add Rx as dependancy and implement base support code
RxTransformer will be used to simplify a bit the code by
using the method compose(). Besides, we'll add as a side
effect the logging of any error that might occur in the flows.

On the other hand, we've added a private compositeDisposable
to the BaseViewModel class. It will be cleared when VM dies